Moved mock/netns to package (Tests ok)
