+++ network/native_endpointclient_linux_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/Azure/azure-container-networking/netio"
 	"github.com/Azure/azure-container-networking/netlink"
+	"github.com/Azure/azure-container-networking/netns"
 	"github.com/Azure/azure-container-networking/network/networkutils"
 	"github.com/Azure/azure-container-networking/platform"
 	"github.com/stretchr/testify/require"
@@ -34,7 +35,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(2, "no such file or directory"),
+				netnsClient:       netns.NewMockNetns(2, "no such file or directory"),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -51,7 +52,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -68,7 +69,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(true, "netlink fail"),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -86,7 +87,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -104,7 +105,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -122,7 +123,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(1, "netns failure"),
+				netnsClient:       netns.NewMockNetns(1, "netns failure"),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -130,7 +131,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 			},
 			epInfo:     &EndpointInfo{},
 			wantErr:    true,
-			wantErrMsg: "NativeEndpointClient Error : Failed to get VM NS handle : " + ErrorMockNetns.Error() + " : netns failure",
+			wantErrMsg: "NativeEndpointClient Error : Failed to get VM NS handle : " + netns.ErrorMockNetns.Error() + " : netns failure",
 		},
 		{
 			name: "Add endpoints NetNS GetFromName fail (with error other than file does not exists)",
@@ -140,7 +141,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(2, "netns failure"),
+				netnsClient:       netns.NewMockNetns(2, "netns failure"),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -148,7 +149,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 			},
 			epInfo:     &EndpointInfo{},
 			wantErr:    true,
-			wantErrMsg: "NativeEndpointClient Error : Error other than vnet NS doesn't exist : " + ErrorMockNetns.Error() + " : netns failure",
+			wantErrMsg: "NativeEndpointClient Error : Error other than vnet NS doesn't exist : " + netns.ErrorMockNetns.Error() + " : netns failure",
 		},
 		{
 			name: "Add endpoints NetNS Set fail",
@@ -158,7 +159,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(3, "netns failure"),
+				netnsClient:       netns.NewMockNetns(3, "netns failure"),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -166,7 +167,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 			},
 			epInfo:     &EndpointInfo{},
 			wantErr:    true,
-			wantErrMsg: "NativeEndpointClient Error : Failed to set current NS to VM : " + ErrorMockNetns.Error() + " : netns failure",
+			wantErrMsg: "NativeEndpointClient Error : Failed to set current NS to VM : " + netns.ErrorMockNetns.Error() + " : netns failure",
 		},
 	}
 
@@ -199,7 +200,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -216,7 +217,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -234,7 +235,7 @@ func TestNativeAddEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -278,7 +279,7 @@ func TestNativeDeleteEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -303,7 +304,7 @@ func TestNativeDeleteEndpoints(t *testing.T) {
 				vnetVethName:      "A1veth0",
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
-				netnsClient:       NewMockNetns(5, "netns failure"),
+				netnsClient:       netns.NewMockNetns(5, "netns failure"),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -318,7 +319,7 @@ func TestNativeDeleteEndpoints(t *testing.T) {
 				},
 			},
 			wantErr:    true,
-			wantErrMsg: "NativeEndpointClient Error : Failed to delete namespace : " + ErrorMockNetns.Error() + " : netns failure",
+			wantErrMsg: "NativeEndpointClient Error : Failed to delete namespace : " + netns.ErrorMockNetns.Error() + " : netns failure",
 		},
 	}
 
@@ -358,7 +359,7 @@ func TestNativeConfigureContainerInterfacesAndRoutes(t *testing.T) {
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
 				vnetMac:           vnetMac,
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -383,7 +384,7 @@ func TestNativeConfigureContainerInterfacesAndRoutes(t *testing.T) {
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
 				vnetMac:           vnetMac,
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -416,7 +417,7 @@ func TestNativeConfigureContainerInterfacesAndRoutes(t *testing.T) {
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
 				vnetMac:           vnetMac,
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(true, "netlink fail"),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -442,7 +443,7 @@ func TestNativeConfigureContainerInterfacesAndRoutes(t *testing.T) {
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
 				vnetMac:           vnetMac,
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -489,7 +490,7 @@ func TestNativeConfigureContainerInterfacesAndRoutes(t *testing.T) {
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
 				vnetMac:           vnetMac,
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
@@ -515,7 +516,7 @@ func TestNativeConfigureContainerInterfacesAndRoutes(t *testing.T) {
 				containerVethName: "B1veth0",
 				vnetNSName:        "az_ns_1",
 				vnetMac:           vnetMac,
-				netnsClient:       NewMockNetns(0, ""),
+				netnsClient:       netns.NewMockNetns(0, ""),
 				netlink:           netlink.NewMockNetlink(false, ""),
 				plClient:          platform.NewMockExecClient(false),
 				netUtilsClient:    networkutils.NewNetworkUtils(nl, plc),
